+++ apps/storybook/src/playground/hooks/useEnvFormUrl.tsx
@@ -8,6 +8,7 @@ export type SearchParams = {
   env?: string;
   theme?: string;
   usePrivy?: boolean;
+  asyncLoadLocale?: boolean;
 };
 
 export function useEnvFormUrl() {
@@ -20,6 +21,7 @@ export function useEnvFormUrl() {
     const brokerId = searchParams.get("brokerId") || undefined;
     const brokerName = searchParams.get("brokerName") || undefined;
     const usePrivy = searchParams.get("usePrivy") || undefined;
+    const asyncLoadLocale = searchParams.get("asyncLoadLocale") || undefined;
 
     return {
       env,
@@ -27,7 +29,10 @@ export function useEnvFormUrl() {
       brokerId,
       brokerName,
       theme,
+      // default true
       usePrivy: usePrivy !== "false",
+      // default false
+      asyncLoadLocale: asyncLoadLocale === "true",
     };
   }, [searchParams]);
 }