+++ packages/i18n/src/locale/module/notification.ts
@@ -6,14 +6,15 @@ export const notification = {
   "notification.joinNow": "Join now",
   "notification.listing": "New token listing",
   "notification.maintenance": "System maintenance",
-  "notification.maintenanceDuration":
+  "notification.maintenanceDuration.hours":
     "{{hours}} HRs at {{startTimeFormatted}} - {{endTimeFormatted}} (UTC)",
+  "notification.maintenanceDuration.minutes":
+    "{{minutes}} MINs at {{startTimeFormatted}} - {{endTimeFormatted}} (UTC)",
   "notification.recentlyUpdated": "Recently updated",
   "notification.title": "Announcement",
   "notification.centerTitle": "Announcement center",
 
   // Notification related keys
-  "notification.campaignTitle": "Latest Campaign is coming",
   "notification.closeAll": "Close all ({{total}})",
   "notification.delistingTitle": "Token Delisting Notice",
   "notification.generalTitle": "General Update",