+++ packages/ui-transfer/src/components/withdrawForm/withdrawForm.ui.tsx
@@ -38,7 +38,6 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
     currentChain,
     fee,
     settingChain,
-    chainVaultBalance,
     crossChainTrans,
     checkIsBridgeless,
     withdrawTo,
@@ -46,7 +45,6 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
     onSourceTokenChange,
     vaultBalanceList,
     qtyGreaterThanMaxAmount,
-    qtyGreaterThanVault,
   } = props;
 
   const { t } = useTranslation();
@@ -64,6 +62,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
         onChange={props.setToAccountId}
         status={props.toAccountIdInputStatus}
         hintMessage={props.toAccountIdHintMessage}
+        disabled={!props.isLoggedIn}
       />
       <Box my={2}>
         <Text size="xs" intensity={54}>
@@ -92,6 +91,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
             vaultBalanceList={vaultBalanceList}
             testId="oui-testid-withdraw-dialog-quantity-input"
             displayType="vaultBalance"
+            disabled={!props.isLoggedIn}
           />
         </Box>
         <AvailableQuantity
@@ -125,7 +125,11 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
         >
           <TabPanel
             title={t("transfer.web3Wallet.my")}
-            icon={<WalletIcon size={"xs"} name={props.walletName ?? ""} />}
+            icon={
+              props.walletName ? (
+                <WalletIcon size={"xs"} name={props.walletName ?? ""} />
+              ) : undefined
+            }
             value={WithdrawTo.Wallet}
           >
             <ChainSelect
@@ -134,6 +138,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
               onValueChange={props.onChainChange}
               wrongNetwork={props.wrongNetwork}
               loading={settingChain}
+              disabled={!props.isLoggedIn}
             />
             <QuantityInput
               classNames={{
@@ -165,11 +170,7 @@ export const WithdrawForm: React.FC<WithdrawFormProps> = (props) => {
       </Box>
       <WithdrawWarningMessage
         checkIsBridgeless={checkIsBridgeless}
-        chainVaultBalance={chainVaultBalance as number}
-        currentChain={currentChain}
         crossChainTrans={crossChainTrans}
-        tokenName={sourceToken?.symbol as string}
-        qtyGreaterThanVault={qtyGreaterThanVault}
         qtyGreaterThanMaxAmount={qtyGreaterThanMaxAmount}
         message={props.warningMessage}
       />