+++ packages/net/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @orderly.network/net
 
+## 2.8.4
+
+### Patch Changes
+
+- publish
+- 32b6043: publish
+
+## 2.8.4-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.8.3
 
 ### Patch Changes