+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,18 @@
 # tsconfig
 
+## 0.11.4
+
+### Patch Changes
+
+- publish
+- 32b6043: publish
+
+## 0.11.4-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.11.3
 
 ### Patch Changes