+++ packages/ui-connector/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @orderly.network/ui-connector
 
+## 2.8.4
+
+### Patch Changes
+
+- publish
+- 32b6043: publish
+- Updated dependencies
+- Updated dependencies [32b6043]
+  - @orderly.network/react-app@2.8.4
+  - @orderly.network/hooks@2.8.4
+  - @orderly.network/i18n@2.8.4
+  - @orderly.network/types@2.8.4
+  - @orderly.network/ui@2.8.4
+  - @orderly.network/ui-chain-selector@2.8.4
+
+## 2.8.4-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.8.4-alpha.0
+  - @orderly.network/hooks@2.8.4-alpha.0
+  - @orderly.network/i18n@2.8.4-alpha.0
+  - @orderly.network/types@2.8.4-alpha.0
+  - @orderly.network/ui@2.8.4-alpha.0
+  - @orderly.network/ui-chain-selector@2.8.4-alpha.0
+
 ## 2.8.3
 
 ### Patch Changes